feat(profile): add ProfileAchievements component for user profile page
