+++ apps/web/src/app/[lang]/[username]/_components/profile-banner/profile-banner-edit-images.tsx
@@ -11,6 +11,7 @@ type ProfileBannerEditImagesProps = {
   selectedItem: SelectedItem
   handleCloseDialog: () => void
 }
+
 export const ProfileBannerEditImages = ({
   selectedItem,
   handleCloseDialog,
