+++ apps/web/src/app/[lang]/[username]/page.tsx
@@ -1,27 +1,21 @@
+import { redirect } from 'next/navigation'
+
 import { PageProps } from '@/types/languages'
-import { ProBadge } from '@/components/pro-badge'
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
-import { Lock, LucideIcon, Pencil } from 'lucide-react'
+import { Lock, Pencil } from 'lucide-react'
 import { getProfileByUsername } from '@/services/api/profiles'
+
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
+import { Button } from '@/components/ui/button'
+
 import { ProfileReviews } from './_components/profile-reviews'
 import { ProfileLists } from './_components/profile-lists'
 import { ProfileBanner } from './_components/profile-banner'
-
-import * as icons from 'lucide-react'
-import { redirect } from 'next/navigation'
-import { getDictionary } from '@/utils/dictionaries'
-import { Button } from '@/components/ui/button'
 import { ProfileForm } from './_components/profile-form'
+import { ProfileImage } from './_components/profile-image'
+import { ProBadge } from '@/components/pro-badge'
 
-const RandomIcon = () => {
-  const iconNames = Object.keys(icons)
-  const randomIconName = iconNames[Math.floor(Math.random() * iconNames.length)]
-  const IconComponent = icons[
-    randomIconName as keyof typeof icons
-  ] as LucideIcon
-
-  return <IconComponent size={16} />
-}
+import { getDictionary } from '@/utils/dictionaries'
+import { ProfileAchievements } from './_components/profile-achievements'
 
 type UserPageProps = PageProps<Record<'username', string>>
 
@@ -46,9 +40,7 @@ const UserPage = async ({ params: { username, lang } }: UserPageProps) => {
         <div className="grid grid-cols-1 gap-0 space-y-8 p-4 lg:grid-cols-3 lg:gap-16 lg:px-16">
           <aside className="col-span-1 -mt-20 flex flex-col space-y-4">
             <div className="space-y-4">
-              <div className="relative z-50 flex aspect-square w-32 items-center justify-center rounded-full border bg-muted text-2xl">
-                {profile.username[0].toUpperCase()}
-              </div>
+              <ProfileImage profile={profile} />
 
               <div className="space-y-2">
                 <div className="flex items-center gap-2">
@@ -70,30 +62,9 @@ const UserPage = async ({ params: { username, lang } }: UserPageProps) => {
                   />
                 </div>
               </div>
-
-              <div className="relative space-y-2 rounded-lg border p-4">
-                <div className="absolute right-0 top-0 flex h-full w-full items-center justify-center rounded-lg border border-dashed bg-black/10 backdrop-blur-[1.5px] dark:bg-black/50">
-                  🚧 {dictionary.profile.work_in_progress}
-                </div>
-
-                <h3 className="font-semibold">
-                  {dictionary.profile.achievements}
-                </h3>
-
-                <div className="grid grid-cols-5 gap-2">
-                  {Array.from({ length: 15 }).map((_, index) => {
-                    return (
-                      <div
-                        className="flex aspect-square items-center justify-center rounded-lg bg-muted"
-                        key={index}
-                      >
-                        <RandomIcon />
-                      </div>
-                    )
-                  })}
-                </div>
-              </div>
             </div>
+
+            <ProfileAchievements dictionary={dictionary} />
           </aside>
 
           <section className="col-span-2">
