+++ apps/web/src/components/item-hover-card/index.ts
@@ -1,0 +1,1 @@
+export * from './item-hover-card'
