+++ apps/web/src/components/list-card/list-card.tsx
@@ -3,6 +3,7 @@
 import { useMemo, useState } from 'react'
 import Link from 'next/link'
 import { toast } from 'sonner'
+import Image from 'next/image'
 import { MoreVertical, Trash } from 'lucide-react'
 
 import { Button } from '@/components/ui/button'
@@ -29,9 +30,7 @@ import { useLanguage } from '@/context/language'
 
 import { List } from '@/types/supabase/lists'
 import { useAuth } from '@/context/auth'
-import { Poster } from '../poster'
-import { cn } from '@/lib/utils'
-import Image from 'next/image'
+
 import { tmdbImage } from '@/utils/tmdb/image'
 
 type ListCardProps = { list: List }
